Move exception class to impl
