+++ src/main/java/org/twonote/rgwadmin4j/impl/ErrorUtils.java
@@ -3,7 +3,6 @@ package org.twonote.rgwadmin4j.impl;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 import okhttp3.Response;
-import org.twonote.rgwadmin4j.RgwAdminException;
 
 import java.lang.reflect.Type;
 import java.util.Map;
